+++ docs/changes.rst
@@ -32,6 +32,8 @@ Pending
 * Added check to prevent the toolbar from being installed when tests
   are running.
 * Added test to example app and command to run the example app's tests.
+* Added warning async is not supported
+
 
 4.3.0 (2024-02-01)
 ------------------